requirements.txt
feedparser
pandas
pytest
python-frontmatter
pyvips
rich
ruamel.yaml
toml
yamale
